% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aova2.R
\name{aova2}
\alias{aova2}
\title{aova2}
\usage{
aova2(formula, data)
}
\arguments{
\item{formula, data}{}
}
\value{
ci plot, anova
}
\description{
two way anova (no repetition / unknown value)
}
\examples{
a<-as.factor(c(rep(1,3),rep(2,3),rep(3,3),rep(4,3)))
b<-c(rep(c(1,2,3),4))
b<-factor(b)
y<-c(NA,97.3,96.7,98.6,98.2,96.9,99,98,97.9,98,97.7,96.5)
z<-data.frame(a,b,y)

}
