% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enova1.R
\name{enova1}
\alias{enova1}
\title{enova1}
\usage{
enova1(formula, data, alpha, na.rm = TRUE, verdose = TRUE)
}
\arguments{
\item{formula, data, alpha, na.rm}{}
}
\value{
CI
}
\description{
one way anova (variance of error)
}
\examples{
y1<-c(8.44,8.36,8.28)
y2<-c(8.59,8.91,8.6)
y3<-c(9.34,9.41,9.69)
y4<-c(8.92,8.92,8.74)
y <- c(y1, y2, y3,y4)
n <- rep(3, 4)
group <- rep(1:4, n)
group<-as.factor(group)
z<-data.frame(y,group)

}
