% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kccd.R
\name{kccd}
\alias{kccd}
\title{kccd}
\usage{
kccd(basis,generators,block,n0,alpha,wbreps,bbreps,randomize,inscribed,coding,oneblock)
}
\arguments{
\item{basis,generators,block,n0,alpha,wbreps,bbreps,randomize,inscribed,coding,oneblock}{}
}
\value{
std.order, run.order, level of factor
}
\description{
central composite design
}
\examples{
basis=3
randomize=FALSE
oneblock=TRUE
}
