% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/munova3.R
\name{munova3}
\alias{munova3}
\title{munova3}
\usage{
munova3(formula, data)
}
\arguments{
\item{formula, data, alpha}{}
}
\value{
ci, plot, anova
}
\description{
three way anova (no repetition / unknown value)
}
\examples{
a<-as.factor(c(rep(1,9),rep(2,9),rep(3,9)))
b<-as.factor(c(1,1,1,2,2,2,3,3,3,1,1,1,2,2,2,3,3,3,1,1,1,2,2,2,3,3,3))
c<-as.factor(rep(c(1,2,3),9))
y<-c(74,86,76,72,91,87,48,65,56,61,NA,71,62,81,77,55,72,63,50,70,60,49,68,64,52,69,60)
z<-data.frame(a,b,c,y)

}
