% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qual.R
\name{qual}
\alias{qual}
\title{qualitative data (one way)}
\usage{
qual(A, B, Y, r, i, j)
}
\arguments{
\item{A, B, Y, r, i, j}{abcd}
}
\value{
anova
}
\description{
development of concepts which help us to understand social phenomena in natural settings, giving due emphasis to the meanings, experiences and views of the participants
}
\examples{
A<-c(1,1,2,2,3,3,4,4)
B<-rep(c("plus","minus"),4)
Y<-c(190.00,10.00,178.00,22.00,194.00,6.00,170.00,30.00)
k<-data.frame(A,B,Y)
k
qual(A,B,Y,200,2,4)

}
