% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosplit.R
\name{sp.plot1}
\alias{sp.plot1}
\title{split(two way)}
\usage{
sp.plot1(block, pplot, splot, aplot, Y)
}
\arguments{
\item{block, pplot, splot, aplot, Y}{abcd}
}
\value{
anova
}
\description{
split of factor
}
\examples{
block<-c(rep(1,18),rep(2,18))
A<-c(rep(1,6),rep(2,6),rep(3,6),rep(1,6),rep(2,6),rep(3,6))
B<-c(rep(1,3),rep(2,3),rep(1,3),rep(2,3),rep(1,3),rep(2,3),rep(1,3),rep(2,3),rep(1,3),rep(2,3),rep(1,3),rep(2,3))
C<-c(rep(c(1,2,3),12))
yield<-c(1,3,-1,-3,4,1,18,21,19,21,16,19,20,18,23,21,22,20,2,0,1,-2,3,0,17,20,19,18,21,16,20,20,21,16,19,20)
k<-data.frame(block,A,B,C,yield,stringsAsFactors = FALSE)
k
x<-with(k,sp.plot1(block,A,B,C,yield))

}
