% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fullynested.R
\name{sp.plot2}
\alias{sp.plot2}
\title{Fully Nested}
\usage{
sp.plot2(pplot, splot, aplot, rplot, y)
}
\arguments{
\item{pplot, splot, aplot, rplot, y}{abcd}
}
\value{
analysis of variance
}
\description{
If all arguments are stake
}
\examples{
A<-rep(c(1,1,2,2,3,3,4,4),4)
B<-c(rep(1,16),rep(2,16))
C<-c(rep(1,8),rep(2,8),rep(1,8),rep(2,8))
R<-c(rep(c(1,2),16))
con<-c(55.30,55.33,55.89,55.82,55.35,55.39,55.30,55.38,55.53,55.55,56.14,56.12,55.59,55.53,55.44,55.45,55.04,55.05,55.56,55.54,55.10,55.06,55.03,54.94,55.22,55.20,55.76,55.84,55.29,55.34,55.12,55.15)
sp.plot2(A,B,C,R,con)

}
