% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notsplit.R
\name{sp.plot3}
\alias{sp.plot3}
\title{can not split}
\usage{
sp.plot3(pplot, splot, y)
}
\arguments{
\item{pplot, splot, y}{abcd}
}
\value{
anova
}
\description{
the factor can not split
}
\examples{
A<-rep(c(1,1,2,2,3,3),3)
B<-c(rep(1,6),rep(2,6),rep(3,6))
y<-c(61.0,60.2,64.1,63.2,65.2,66.1,63.3,62.7,66.2,65.4,66.6,67.2,61.3,61.9,63.2,64.2,66.0,66.4)
sp.plot3(A,B,y)

}
